Renamed tutorial examples.
