+++ lys_instr/tutorial/AdvEx1.py
@@ -42,7 +42,6 @@ class window(LysSubWindow):
 
         w = QtWidgets.QWidget()
         w.setLayout(HBox)
-        self.setWidget(w)       # cannot set layout
+        self.setWidget(w)
 
-        # self.adjustSize()
